+++ tests/path.test.php
@@ -4,4 +4,6 @@ require 'vendor/autoload.php';
 
 use Path\Path;
 
+
+
 ?>